requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
mangum==0.17.0
pillow==10.4.0
numpy==1.26.4
scikit-learn==1.4.2
opencv-python-headless==4.9.0.80
joblib==1.3.2